+++ engine/development/deploymentManager/src/main/scala/pl/touk/nussknacker/development/manager/MockableDeploymentManagerProvider.scala
@@ -12,7 +12,7 @@ import pl.touk.nussknacker.engine.api.definition.{NotBlankParameterValidator, St
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.deployment.simple.{SimpleProcessStateDefinitionManager, SimpleStateStatus}
 import pl.touk.nussknacker.engine.api.process.{ProcessIdWithName, ProcessName, VersionId}
-import pl.touk.nussknacker.engine.deployment.{CustomActionDefinition, CustomActionParameter, ExternalDeploymentId}
+import pl.touk.nussknacker.engine.deployment.ExternalDeploymentId
 import pl.touk.nussknacker.engine.management.{FlinkProcessTestRunner, FlinkStreamingPropertiesConfig}
 import pl.touk.nussknacker.engine.newdeployment.DeploymentId
 import pl.touk.nussknacker.engine.testing.StubbingCommands
@@ -81,35 +81,6 @@ object MockableDeploymentManagerProvider {
 
     override def processStateDefinitionManager: ProcessStateDefinitionManager =
       SimpleProcessStateDefinitionManager
-
-    override def customActionsDefinitions: List[CustomActionDefinition] = {
-      import SimpleStateStatus._
-      List(
-        deployment.CustomActionDefinition(
-          actionName = ScenarioActionName("hello"),
-          allowedStateStatusNames = List(ProblemStateStatus.name, NotDeployed.name)
-        ),
-        deployment.CustomActionDefinition(
-          actionName = ScenarioActionName("not-implemented"),
-          allowedStateStatusNames = List(ProblemStateStatus.name, NotDeployed.name)
-        ),
-        deployment.CustomActionDefinition(
-          actionName = ScenarioActionName("some-params-action"),
-          allowedStateStatusNames = List(ProblemStateStatus.name, NotDeployed.name),
-          parameters = List(
-            CustomActionParameter(
-              "param1",
-              StringParameterEditor,
-              NotBlankParameterValidator :: Nil
-            )
-          )
-        ),
-        deployment.CustomActionDefinition(
-          actionName = ScenarioActionName("invalid-status"),
-          allowedStateStatusNames = Nil
-        )
-      )
-    }
 
     override def getProcessStates(name: ProcessName)(
         implicit freshnessPolicy: DataFreshnessPolicy